Refactor lib folders
