+++ libs/types/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/types';
+export * from './types';
