+++ libs/ui-lib/src/components/custom-text-input/custom-text-input.scss
