+++ libs/ui-lib/src/components/custom-time-picker/custom-time-picker.scss
