+++ libs/ui-lib/src/index.ts
@@ -1,4 +1,3 @@
-export * from './lib/custom-time-picker/custom-time-picker';
-export * from './lib/custom-text-input/custom-text-input';
-export * from './lib/pill-options/pill-options';
-export * from './lib/ui-lib';
+export * from './components/custom-time-picker/custom-time-picker';
+export * from './components/custom-text-input/custom-text-input';
+export * from './components/pill-options/pill-options';
