--- libs/ui-lib/src/lib/ui-lib.spec.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-
-import UiLib from './ui-lib';
-
-describe('UiLib', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<UiLib />);
-    expect(baseElement).toBeTruthy();
-  });
-});
