--- libs/ui-lib/src/lib/ui-lib.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import './ui-lib.scss';
-
-/* eslint-disable-next-line */
-export interface UiLibProps {}
-
-export const UiLib = (props: UiLibProps) => {
-  return (
-    <div>
-      <h1>Welcome to ui-lib!</h1>
-    </div>
-  );
-};
-
-export default UiLib;
