+++ libs/ui-lib/src/styles/styles.scss
