+++ workspace.json
@@ -11,7 +11,7 @@
           "builder": "@nrwl/web:build",
           "options": {
             "stylePreprocessorOptions": {
-              "includePaths": ["libs/ui-lib/src/lib/styles"]
+              "includePaths": ["libs/ui-lib/src/styles"]
             },
             "extractCss": true,
             "outputPath": "dist/apps/mull-ui",
@@ -24,7 +24,7 @@
               "apps/mull-ui/src/manifest.json",
               "apps/mull-ui/src/assets"
             ],
-            "styles": ["libs/ui-lib/src/lib/styles/styles.scss"],
+            "styles": ["libs/ui-lib/src/styles/styles.scss"],
             "scripts": [],
             "webpackConfig": "@nrwl/react/plugins/webpack"
           },
